backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==0.8.11
regex==2023.10.3
python-dateutil==2.8.2
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0 